@@ -20,7 +20,6 @@ import (
 	"encoding/json"
 	"errors"
 	"io/fs"
-	"io/ioutil"
 	"log"
 	"os"
 
@@ -33,7 +32,7 @@ func main() {
 		log.Fatal("version not found")
 	}
 
-	schemaContents, err := ioutil.ReadFile("./schema.json")
+	schemaContents, err := os.ReadFile("./schema.json")
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -56,7 +55,7 @@ func main() {
 		log.Fatal(err)
 	}
 
-	err = ioutil.WriteFile("./schema-embed.json", versionedContents, 0600)
+	err = os.WriteFile("./schema-embed.json", versionedContents, 0600)
 	if err != nil {
 		log.Fatal(err)
 	}
